adc fs

+++ inc/filters.h
@@ -9,8 +9,8 @@
 #define FILTERS_H_
 
 extern const float h[N_RAW];
-extern uint16_t taps[N_RAW];
+extern float taps[N_RAW];
 
-float filter(uint16_t sample, const float* h, uint16_t* taps, uint8_t length);
+float filter(float sample, const float* h, float* taps, uint8_t length);
 
 #endif /* FILTERS_H_ */
